Update tests to use new interface_py package
